compute dataset suitable for display using flot
Yak is structured as a list of commits giving counters for various phases.
To display CFDs in flot, a more convenient format is having one series of
data per phase, where each series is made of points of the form (time, counter).

This commit provides an inefficient transformation from a Yak to such a list,
along with needed JSON transformations. HTML display does not quite work yet.